requirements.txt
discord.py
requests==2.23.0
pprint==0.1
